terminando categoria teste

+++ backend/src/core/use-cases/usuario/UsuarioUseCase.ts
@@ -18,7 +18,6 @@ export class CriarUsuarioUseCase {
     const usuario = new Usuario({user_id: user_id, nome: nome, foto_url: foto_url});
     await this.usuarioRepository.create(usuario);
  
-    console.log('criei um usuario: ', usuario);
     return usuario;
   }
 
@@ -28,8 +27,13 @@ export class DeletarUsuarioUseCase {
   constructor(private usuarioRepository: IUsuarioRepository){}
 
   async executar(user_id: string): Promise<boolean>{
-    const existente = await this.usuarioRepository.findByUserId(user_id)
-      if (!existente) throw new ResourceNotFoundError('Usuario');
+    validarUUID(user_id, 'ID do usuário');
+    const usuario = await this.usuarioRepository.findByUserId(user_id)
+
+    if (!usuario) {
+      throw new ResourceNotFoundError('Usuario');
+    }
+
     await this.usuarioRepository.delete(user_id);
     return true;
   }
@@ -47,7 +51,6 @@ export class AtualizarUsuarioUseCase {
     }
     const usuarioInstanciado = new Usuario(usuario);
     usuarioInstanciado.atualizarPerfil(dados); 
-    console.log('usuario atulizado:', usuarioInstanciado);
     await this.usuarioRepository.update(usuarioInstanciado);
     return usuarioInstanciado;
   }
@@ -64,7 +67,6 @@ export class PesquisarPorUserId {
     if (!usuario) {
       throw new ResourceNotFoundError('Usuário');
     }
-    console.log('busquei um ususario: ', usuario);
     return usuario;
   }
 }
